+++ RingSoft.DevLogix.Library/ViewModels/CustomerManagement/SupportTicketCostRow.cs
@@ -47,7 +47,7 @@ namespace RingSoft.DevLogix.Library.ViewModels.CustomerManagement
                     {
                         FormatType = DecimalEditFormatTypes.Currency,
                         Precision = 2,
-                    }, (decimal)Cost);
+                    }, Cost);
                 default:
                     throw new ArgumentOutOfRangeException();
             }
