+++ RingSoft.DevLogix.Library/ViewModels/ProjectManagement/Project/ProjectUsersGridRow.cs
@@ -62,7 +62,7 @@ namespace RingSoft.DevLogix.Library.ViewModels.ProjectManagement
                     return new DataEntryGridDecimalCellProps(this, columnId, new DecimalEditControlSetup
                     {
                         FormatType = DecimalEditFormatTypes.Currency,
-                    }, (decimal)Cost);
+                    }, Cost);
                 case ProjectUserColumns.IsStandard:
                     return new DataEntryGridCheckBoxCellProps(this, columnId, IsStandard);
                 case ProjectUserColumns.Sunday:
