+++ RingSoft.DevLogix.Library/ViewModels/QualityAssurance/ErrorUserRow.cs
@@ -45,7 +45,7 @@ namespace RingSoft.DevLogix.Library.ViewModels.QualityAssurance
                     {
                         FormatType = DecimalEditFormatTypes.Currency,
                         Precision = 2,
-                    }, (decimal)Cost);
+                    }, Cost);
                 default:
                     throw new ArgumentOutOfRangeException();
             }
