+++ RingSoft.DevLogix.Library/ViewModels/QualityAssurance/Testing/TestingOutlineCostRow.cs
@@ -45,7 +45,7 @@ namespace RingSoft.DevLogix.Library.ViewModels.QualityAssurance.Testing
                     {
                         FormatType = DecimalEditFormatTypes.Currency,
                         Precision = 2,
-                    }, (decimal)Cost);
+                    }, Cost);
                 default:
                     throw new ArgumentOutOfRangeException();
             }
