+++ RingSoft.DevLogix.Library/ViewModels/UserManagement/UserMaintenanceViewModel.cs
@@ -1290,10 +1290,10 @@ namespace RingSoft.DevLogix.Library.ViewModels.UserManagement
                     e.HeaderRow.StringField16 = AppGlobals.MakeTimeSpent(user.NonBillableProjectsMinutesSpent);
                     e.HeaderRow.StringField17 = AppGlobals.MakeTimeSpent(user.ErrorsMinutesSpent);
                     e.HeaderRow.StringField18 = AppGlobals.MakeTimeSpent(user.TestingOutlinesMinutesSpent);
-                    e.HeaderRow.NumberField15 = percentSetup.FormatValue((decimal)billabilityData.BillableProjects);
-                    e.HeaderRow.NumberField16 = percentSetup.FormatValue((decimal)billabilityData.NonBillableProjects);
-                    e.HeaderRow.NumberField17 = percentSetup.FormatValue((decimal)billabilityData.Errors);
-                    e.HeaderRow.NumberField18 = percentSetup.FormatValue((decimal)billabilityData.TestingOutlines);
+                    e.HeaderRow.NumberField15 = percentSetup.FormatValue(billabilityData.BillableProjects);
+                    e.HeaderRow.NumberField16 = percentSetup.FormatValue(billabilityData.NonBillableProjects);
+                    e.HeaderRow.NumberField17 = percentSetup.FormatValue(billabilityData.Errors);
+                    e.HeaderRow.NumberField18 = percentSetup.FormatValue(billabilityData.TestingOutlines);
 
                     foreach (var userTimeOff in user.UserTimeOff)
                     {
